Moved worker_class.py to libE_worker.py
