+++ libensemble/tests/unit_tests/tofix__test_worker.py
@@ -1,6 +1,6 @@
 import libensemble.tests.unit_tests.setup as setup
 from libensemble.libE_fields import libE_fields
-from libensemble.worker_class import Worker
+from libensemble.libE_worker import Worker
 from libensemble.message_numbers import EVAL_SIM_TAG, EVAL_GEN_TAG
 import numpy as np
